+++ app/src/main/java/com/example/taskapp/ui/app/DoingFragment.kt
@@ -5,13 +5,18 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.taskapp.data.model.Status
+import com.example.taskapp.data.model.Task
 import com.example.taskapp.databinding.FragmentDoingBinding
+import com.example.taskapp.ui.adapter.TaskAdapter
 
 
 class DoingFragment : Fragment() {
     private var _binding: FragmentDoingBinding? = null
     private val binding get() = _binding!!
 
+    lateinit var taskAdapter: TaskAdapter
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -19,6 +24,43 @@ class DoingFragment : Fragment() {
         _binding = FragmentDoingBinding.inflate(inflater, container, false)
         return binding.root
     }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initRecyclerViewTask(getTasks())
+    }
+
+    private fun initRecyclerViewTask(taskList:List<Task>){
+        taskAdapter = TaskAdapter(taskList)
+
+        binding.rvTasks.layoutManager = LinearLayoutManager(requireContext())
+        binding.rvTasks.setHasFixedSize(true)
+        binding.rvTasks.adapter=taskAdapter
+    }
+
+    private  fun getTasks()= listOf(
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+        Task("0","Criar nova tela do app", Status.TODO),
+    )
 
 
     override fun onDestroyView() {