+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- RecyclerView for tasks
 - Task Card style
 
 ## [0.4.0] - 2023.11.17